+++ integration_test.go
@@ -131,7 +131,7 @@ func TestIntegration_PGDUMP_SimpleUse(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to open the test data file: %v", err)
 	}
-	err = toSpanner(conversion.PGDUMP, projectID, instanceID, dbName, &conversion.IOStreams{In: f, Out: os.Stdout}, filePrefix, now)
+	err = commandLine(conversion.PGDUMP, projectID, instanceID, dbName, &conversion.IOStreams{In: f, Out: os.Stdout}, filePrefix, now)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -188,7 +188,7 @@ func TestIntegration_POSTGRES_SimpleUse(t *testing.T) {
 	dbPath := fmt.Sprintf("projects/%s/instances/%s/databases/%s", projectID, instanceID, dbName)
 	filePrefix = filepath.Join(tmpdir, dbName+".")
 
-	err := toSpanner(conversion.POSTGRES, projectID, instanceID, dbName, &conversion.IOStreams{Out: os.Stdout}, filePrefix, now)
+	err := commandLine(conversion.POSTGRES, projectID, instanceID, dbName, &conversion.IOStreams{Out: os.Stdout}, filePrefix, now)
 	if err != nil {
 		t.Fatal(err)
 	}
