+++ web/postgres_typemap.go
@@ -22,15 +22,15 @@ import (
 var postgresTypeMap = map[string][]typeIssue{
 	"bool": []typeIssue{
 		typeIssue{T: ddl.Bool},
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"boolean": []typeIssue{
 		typeIssue{T: ddl.Bool},
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"bigserial": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Serial}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Serial].Brief}},
 	"bpchar": []typeIssue{
 		typeIssue{T: ddl.String},
 		typeIssue{T: ddl.Bytes}},
@@ -38,75 +38,75 @@ var postgresTypeMap = map[string][]typeIssue{
 		typeIssue{T: ddl.String},
 		typeIssue{T: ddl.Bytes}},
 	"bytea": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Bytes}},
 	"date": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Date}},
 	"float8": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Float64}},
 	"double precision": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Float64}},
 	"float4": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Float64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Float64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"real": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Float64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Float64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"int8": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Int64}},
 	"bigint": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Int64}},
 	"int4": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"integer": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"int2": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"smallint": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Widened}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"numeric": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Float64, Issue: internal.Numeric}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Float64, Brief: internal.IssueDB[internal.Numeric].Brief}},
 	"serial": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Int64, Issue: internal.Serial}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Int64, Brief: internal.IssueDB[internal.Serial].Brief}},
 	"text": []typeIssue{
-		typeIssue{T: ddl.Bytes, Issue: internal.Widened},
+		typeIssue{T: ddl.Bytes, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.String}},
 	"timestamptz": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Timestamp}},
 	"timestamp with time zone": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
 		typeIssue{T: ddl.Timestamp}},
 	"timestamp": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Timestamp, Issue: internal.Timestamp}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Timestamp, Brief: internal.IssueDB[internal.Timestamp].Brief}},
 	"timestamp without time zone": []typeIssue{
-		typeIssue{T: ddl.String, Issue: internal.Widened},
-		typeIssue{T: ddl.Timestamp, Issue: internal.Timestamp}},
+		typeIssue{T: ddl.String, Brief: internal.IssueDB[internal.Widened].Brief},
+		typeIssue{T: ddl.Timestamp, Brief: internal.IssueDB[internal.Timestamp].Brief}},
 	"varchar": []typeIssue{
 		typeIssue{T: ddl.String},
-		typeIssue{T: ddl.Bytes, Issue: internal.Widened}},
+		typeIssue{T: ddl.Bytes, Brief: internal.IssueDB[internal.Widened].Brief}},
 	"character varying": []typeIssue{
 		typeIssue{T: ddl.String},
-		typeIssue{T: ddl.Bytes, Issue: internal.Widened}},
+		typeIssue{T: ddl.Bytes, Brief: internal.IssueDB[internal.Widened].Brief}},
 }
 
-func toSpannerTypePostgres(conv *internal.Conv, id string, toType string, mods []int64) (ddl.Type, []internal.SchemaIssue) {
+func toSpannerTypePostgres(srcType string, spType string, mods []int64) (ddl.Type, []internal.SchemaIssue) {
 
-	switch id {
+	switch srcType {
 	case "bool", "boolean":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		case ddl.Int64:
@@ -116,14 +116,14 @@ func toSpannerTypePostgres(conv *internal.Conv, id string, toType string, mods [
 		}
 
 	case "bigserial":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Int64}, []internal.SchemaIssue{internal.Serial}
 		}
 	case "bpchar", "character": // Note: Postgres internal name for char is bpchar (aka blank padded char).
-		switch toType {
+		switch spType {
 		case ddl.Bytes:
 			if len(mods) > 0 {
 				return ddl.Type{Name: ddl.Bytes, Len: mods[0]}, nil
@@ -137,56 +137,56 @@ func toSpannerTypePostgres(conv *internal.Conv, id string, toType string, mods [
 			return ddl.Type{Name: ddl.String, Len: 1}, nil
 		}
 	case "bytea":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Bytes, Len: ddl.MaxLength}, nil
 		}
 	case "date":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Date}, nil
 		}
 	case "float8", "double precision":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Float64}, nil
 		}
 	case "float4", "real":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Float64}, []internal.SchemaIssue{internal.Widened}
 		}
 	case "int8", "bigint":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Int64}, nil
 		}
 	case "int4", "integer":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Int64}, []internal.SchemaIssue{internal.Widened}
 		}
 	case "int2", "smallint":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Int64}, []internal.SchemaIssue{internal.Widened}
 		}
 	case "numeric": // Map all numeric types to float64.
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
@@ -199,21 +199,21 @@ func toSpannerTypePostgres(conv *internal.Conv, id string, toType string, mods [
 			return ddl.Type{Name: ddl.Float64}, []internal.SchemaIssue{internal.Numeric}
 		}
 	case "serial":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Int64}, []internal.SchemaIssue{internal.Serial}
 		}
 	case "text":
-		switch toType {
+		switch spType {
 		case ddl.Bytes:
 			return ddl.Type{Name: ddl.Bytes, Len: ddl.MaxLength}, nil
 		default:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, nil
 		}
 	case "timestamptz", "timestamp with time zone":
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
@@ -221,14 +221,14 @@ func toSpannerTypePostgres(conv *internal.Conv, id string, toType string, mods [
 		}
 	case "timestamp", "timestamp without time zone":
 		// Map timestamp without timezone to Spanner timestamp.
-		switch toType {
+		switch spType {
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
 			return ddl.Type{Name: ddl.Timestamp}, []internal.SchemaIssue{internal.Timestamp}
 		}
 	case "varchar", "character varying":
-		switch toType {
+		switch spType {
 		case ddl.Bytes:
 			return ddl.Type{Name: ddl.Bytes, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
